Include license info in dependency list
