+++ NOTICE.markdown
@@ -1,13 +1,17 @@
-# List of external packages and their license
+```
+   Copyright 2017 Mesosphere
 
-* finagle: APLv2.0
-* json-schema-validator: APLv2.0 / LGPLv3.0
-* logback-classic: EPLv1.0 / LGPLv2.1
-* finch: APLv2.0
-* mustache.java: APLv2.0
-* scala-uri: APLv2.0
-* circe: APLv2.0
-* finch-server: APLv2.0
-* curator: APLv2.0
-* twitter-util: APLv2.0
-* twitter-bijection: APLv2.0
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+```
+
+[List of external packages with their licenses](project/Deps.scala)
